Fixed the OnSelectedIndexChange to show correct items.
Note database is now functioning properly. HOWEVER,
Doing any Server Explorer operations (opening the database, seeing the tables) OR doing any Data Sources operations (Seeing the DataSet, etc) will delete the data in the program and

WILL ADD AN EXTRA FILE in the Dataset folder with a "1" at the end. This will cause a large number of errors. DELETE THIS FILE WITH A "1" IN THE FILE NAME TO EXECUTE PROPERLY.